requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
arabic-reshaper==2.1.3
arrow==1.2.0; python_version >= '3.6'
asgiref==3.4.1; python_version >= '3.6'
blessed==1.19.0; python_version >= '2.7'
cachetools==4.2.4; python_version ~= '3.5'
certifi==2021.5.30
charset-normalizer==2.0.6; python_version >= '3'
django-appconf==1.0.5; python_version >= '3.6'
django-cors-headers==3.10.0
django-extensions==3.1.3
django-filter==21.1
django-googledrive-storage==1.5.2
django-js-asset==2.2.0
django-mptt
django-picklefield==3.0.1
django-q==1.3.9
django-querycount==0.8.3
django-redis==5.4.0
django-silk==5.0.3
django-tinymce==3.3.0
django-userforeignkey==0.3.0
django==3.2.8
djangorestframework==3.12.4
docopt==0.6.2
#django-mptt==0.14.0
et-xmlfile==1.1.0; python_version >= '3.6'
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-api-core==2.1.0; python_version >= '3.6'
google-api-python-client==2.24.0; python_version >= '3.6'
google-auth-httplib2==0.1.0
google-auth==2.2.1; python_version >= '3.6'
googleapis-common-protos==1.53.0; python_version >= '3.6'
html5lib==1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
httplib2==0.19.1
icecream==2.1.3
idna==3.2; python_version >= '3'
markdown==3.3.4
mysqlclient==2.0.3
num2words==0.5.10
oauth2client==4.1.3
openpyxl==3.0.9
pillow
protobuf==3.18.1; python_version >= '3.5'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pyjwt==2.1.0
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pypdf2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.19.0
pytz==2021.3
redis==3.5.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
reportlab==3.6.1; python_version >= '2.7' and python_version < '4'
requests==2.26.0
rsa==4.7.2; python_version >= '3.5' and python_version < '4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
uritemplate==3.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
wcwidth==0.2.5
webencodings==0.5.1
xhtml2pdf==0.2.5
qrcode==7.4.2
pdf2image==1.16.3
flake8==6.1.0
slackclient==2.9.4
weasyprint==60.2
